create input component
