+++ src/components/input/input.scss
@@ -1,0 +1,14 @@
+@import "../../style.scss";
+
+div {
+    @include themify($themes) {
+        input {
+            background-color:themed('bgSoft');
+            border: 1px solid themed('border');
+            padding: 20px 10px;
+            width: 100%;
+            border-radius: 5px;;
+        }
+
+}
+}
